backend/requirements.txt
faiss-cpu==1.7.4
numpy>=1.24.3
scikit-learn>=1.2.2
pandas>=2.0.2
python-dotenv>=1.0.0
transformers>=4.29.2
sentence-transformers>=2.2.2
torch>=2.0.1